Rename files
Kept this separate so that hopefully GitHub does a reasonable job of showing a diff, but if not, you can just ignore this commit. There are no content changes.
