+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/ICompletionListProvider.cs
@@ -1,0 +1,22 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using System.Collections.Immutable;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.VisualStudio.LanguageServer.Protocol;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
+{
+    internal interface ICompletionListProvider
+    {
+        public ImmutableHashSet<string> TriggerCharacters { get; }
+
+        public Task<VSInternalCompletionList?> GetCompletionListAsync(
+            int absoluteIndex,
+            VSInternalCompletionContext completionContext,
+            DocumentContext documentContext,
+            VSInternalClientCapabilities clientCapabilities,
+            CancellationToken cancellationToken);
+    }
+}
